refactor: 컴포넌트 common 제거(폴더 구조 평탄화)
