+++ src/features/landingPage/components/BlackSquare.tsx
