+++ src/features/landingPage/components/PurpleCircle.tsx
