--- src/features/landingPage/components/common/FeatureMenu.tsx
@@ -1,12 +1,0 @@
-const FeatureMenu = () => {
-  return (
-    <div className="flex flex-col gap-28 text-[64px] ml-36">
-      <h1>메뉴1</h1>
-      <h1>메뉴2</h1>
-      <h1>메뉴3</h1>
-      <h1>메뉴4</h1>
-    </div>
-  );
-};
-
-export default FeatureMenu;
