--- src/features/landingPage/components/common/LoadLanding.tsx
@@ -1,9 +1,0 @@
-const LoadLanding = () => {
-  return (
-    <div className="fixed top-0 left-0 w-full h-screen flex justify-center items-center bg-white flex-col mt-[-24px]">
-      <h1 className="custom-font text-[194px] max-sm:text-8xl text-black">IT:PLACE</h1>
-    </div>
-  );
-};
-
-export default LoadLanding;
